도시를 두 배로 넓히고, 고속도로에서 차가 떨어지지 않게
도시
- 반지름 112 → 224, 격자선 9줄 → 17줄, 순환도로 104 → 208.
  공항과의 거리도 420 → 640 으로 밀었다
- 자리 고르기를 '조건에 맞는 첫 자리'에서 '점수가 가장 높은 자리'로 바꿨다.
  이 세계는 지면의 6할이 물이라 반지름 224 짜리 원 안에 바다가 절반 넘게
  들어간다. 예전 기준으로는 어느 후보도 통과하지 못해 도시가 아예 안 생겼다
- 도시 원 안의 바다는 바닥까지 메운다 (매립지). 안 메우면 지면 아래가 물로
  남아 물가에서 땅이 떠 보인다
- 도면(ops)을 Int32 → Int16 으로. 값이 모두 16비트에 들어가므로 안전하고,
  세 도시 합쳐 52MB → 26MB
- 구획이 순환도로를 침범하지 않게 반경 기준을 순환도로에 맞췄다
- 넓어진 만큼 버스 정거장 8 → 12곳, 차 38 → 72대

고속도로
- 가드레일을 한 칸 높이 · 군데군데 끊긴 것에서 두 칸 높이 · 빈틈없음으로
  바꿨다. 한 칸짜리는 차가 그냥 타고 넘어 들판·바다로 떨어졌다
- 다리 위에도 난간을 세웠다. 상판 위를 치우는 코드가 가드레일까지 지우고
  있어서 다리 구간만 뻥 뚫려 있었다

검증: 도로 가장자리 232곳 중 두 칸 난간이 없는 곳 1곳(0.4%, 한 칸짜리
틈이라 차 앞모서리 두 곳 중 하나는 반드시 막힌다). 고속도로 위에서 핸들을
좌·우로 끝까지 꺾은 채 45초씩 주행 — 노면을 벗어난 순간 0회, 높이 변화 0.
도시 세 곳 모두 다른 두 도시와 고속도로로 이어지고, 나들목이 순환도로
반지름(208)에 정확히 닿는다. 도시·마을·공항 세 가지 시작 모두 오류 없이 부팅.

+++ web/js/village.js
@@ -101,8 +101,11 @@ VillagePlan.prototype.fill = function (x0, y0, z0, x1, y1, z1, id, meta, force) 
 
 // 배열을 정수 배열로 굳혀 메모리를 아낀다
 VillagePlan.prototype.freeze = function () {
+  // 담기는 값이 모두 16비트에 들어간다 (칸 안 좌표 0~15, y 0~159,
+  // 블록 번호 1000 미만, run 160 이하). 도시가 커지면서 도면이 수십 MB가
+  // 되므로 절반으로 줄인다.
   const self = this;
-  this.ops.forEach(function (a, k) { self.ops.set(k, Int32Array.from(a)); });
+  this.ops.forEach(function (a, k) { self.ops.set(k, Int16Array.from(a)); });
 };
 
 // ── 지형 고르기 ───────────────────────────────────────────────────────
